+++ app/models/concerns/translatable.rb
@@ -12,6 +12,14 @@ module Translatable
     self.slug_ca ||= self.slug
   end
 
+  # Copia el valor de la columna a las traducciones que no tienen nada
+  def propagate_translation(column)
+    value = send(column)
+    [:es, :ca].each do |locale|
+      send("#{column}_#{locale}=", value) unless send("#{column}_#{locale}?")
+    end
+  end
+
   module ClassMethods
     def translates(*columns)
       columns.each do |column|
@@ -27,6 +35,12 @@ module Translatable
         define_method "#{column}=" do |value|
           self.send("#{column}_#{I18n.locale}=", value)
         end
+      end
+    end
+
+    def translation_required(*columns) 
+      before_validation do
+        columns.each {|col| propagate_translation(col) }
       end
     end
   end