+++ app/models/phase.rb
@@ -6,7 +6,9 @@ class Phase < ActiveRecord::Base
   translates :name, :description
   attr_accessible :name_es, :description_es
   attr_accessible :name_ca, :description_ca
-  attr_accessible :name, :description
+  attr_accessible :name, :description, :position
 
   validates_presence_of :name_es, :name_ca
+
+  has_many :categories
 end